Rolled-back variable-array length support on master
The modifications contained bugs I don't have the time
to address right now. Benefits to labCA are small - it
would require additional support in 'glue' and the user-
interface to make this feature accessible to the user.

As-is it would probably also break under older EPICS
versions which lack support for variable-length arrays.